+++ packages/rockets-server/src/typeorm.ts
@@ -1,1 +1,0 @@
-export { defineTypeOrmRepository } from './repository/define-typeorm-repository';